refactor: consolidate InMemoryClipboard implementation into file_clipboard module
