--- src/clipboard/in_memory_clipboard.rs
@@ -1,35 +1,0 @@
-use crate::clipboard::Clipboard;
-use crate::error::AppError;
-use std::cell::RefCell;
-
-#[derive(Default)]
-pub struct InMemoryClipboard {
-    buffer: RefCell<String>,
-}
-
-impl InMemoryClipboard {
-    pub fn with_content(value: &str) -> Self {
-        let cb = Self::default();
-        cb.set_contents(value);
-        cb
-    }
-
-    pub fn contents(&self) -> String {
-        self.buffer.borrow().clone()
-    }
-
-    pub fn set_contents(&self, value: &str) {
-        self.buffer.replace(value.to_string());
-    }
-}
-
-impl Clipboard for InMemoryClipboard {
-    fn copy(&self, text: &str) -> Result<(), AppError> {
-        self.buffer.replace(text.to_string());
-        Ok(())
-    }
-
-    fn paste(&self) -> Result<String, AppError> {
-        Ok(self.buffer.borrow().clone())
-    }
-}
